win9x: warning refs #1  svn merge -r 2753:2757 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/branches/yui/WORK_01
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@2770 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ win9x/menu.cpp
@@ -99,7 +99,7 @@ UINT InsertMenuResource(HMENU hMenu, UINT uItem, BOOL fByPosition, UINT nMenuID)
  */
 static BOOL AppendMenuString(HMENU hMenu, UINT uFlags, UINT_PTR uIDNewItem)
 {
-	std::tstring rString(LoadTString(uIDNewItem));
+	std::tstring rString(LoadTString(static_cast<UINT>(uIDNewItem)));
 	if (rString.empty())
 	{
 		return FALSE;
@@ -127,7 +127,7 @@ static BOOL AppendMenuString(HMENU hMenu, UINT uFlags, UINT_PTR uIDNewItem)
  */
 static BOOL InsertMenuString(HMENU hMenu, UINT uItem, UINT uFlags, UINT_PTR uIDNewItem)
 {
-	std::tstring rString(LoadTString(uIDNewItem));
+	std::tstring rString(LoadTString(static_cast<UINT>(uIDNewItem)));
 	if (rString.empty())
 	{
 		return FALSE;
